Виджеты из PhotoListPage вынесены в папку widgets
