+++ lib/ui/screen/photo_list/widgets/loading_indicator.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/cupertino.dart';
+
+class LoadingIndicator extends StatelessWidget {
+  const LoadingIndicator({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const SliverPadding(
+      padding: EdgeInsets.only(bottom: 25),
+      sliver: SliverToBoxAdapter(
+        child: Center(
+            child: SizedBox(
+              width: 22,
+              height: 22,
+              child: CupertinoActivityIndicator(
+                color: Color.fromRGBO(0, 0, 0, 1),
+              ),
+            )),
+      ),
+    );
+  }
+}
